First Come First Serve (FcFs) Scheduling Algorith
First Come First Served (FCFS) is a Non-Preemptive scheduling algorithm that assigns priority to processes based on the order in which they arrive. The first process to request the CPU is allocated the CPU first, and the subsequent processes are placed in a queue. The CPU works through the queue in the order in which the processes were received, completing one task before moving on to the next. This strategy is often implemented using a FIFO (First In First Out) queue, where new tasks are added to the end of the queue and the CPU removes tasks from the front of the queue. The FCFS algorithm is simple to implement and understand, but it may lead to lower overall system performance if some tasks have longer execution times than others.